Correctly fixed the level transition collider bug.
This bug would occur after the player respawned. It would then not be possible to transition to the next level.
This occured because there was no transition from the respawn animation to the end of level animation. Adding a
transiton between these two animations allowed for the player to progress to the next level even after a
respawn.